Refactor, format checking fix, persist data
