+++ FlutterApp/lib/screens/add.dart
@@ -17,6 +17,8 @@ import 'package:new_ui/screens/mealsuggestions.dart';
 import 'package:path/path.dart' as path;
 import 'package:universal_platform/universal_platform.dart';
 
+import '../components/loaderdialog.dart';
+
 String domain = getDomain(1); //0 IS FOR DEVELOPMENT, 1 IS FOR PRODUCTION
 
 String responseTitle = "";
@@ -24,143 +26,6 @@ String responseText1 = "";
 String responseText2 = "";
 String responseText3 = "";
 String responseColor = "";
-
-//TEMPORARY
-final GlobalKey<State> _LoaderDialog = GlobalKey<State>();
-
-class LoaderDialog {
-  static Future<void> showLoadingDialog(
-      BuildContext context, GlobalKey key) async {
-    //var wid = MediaQuery.of(context).size.width / 2;
-    return showDialog<void>(
-      context: context,
-      //barrierDismissible: false,
-      builder: (BuildContext context) {
-        return Center(
-          child: Padding(
-            padding: const EdgeInsets.only(
-                left: 32, right: 32, top: 80, bottom: 100),
-            child: Material(
-              color: Colors.indigo,
-              elevation: 2,
-              shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(32)),
-              child: SingleChildScrollView(
-                child: Padding(
-                  padding: const EdgeInsets.only(
-                      left: 18, right: 18, top: 16, bottom: 18),
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    //crossAxisAlignment: CrossAxisAlignment.end,
-                    children: [
-                      Padding(
-                        padding: const EdgeInsets.only(
-                            left: 15,
-                            right: 15,
-                            top: 20,
-                            bottom: 20), //apply padding to all four sides
-                        child: Text(responseTitle,
-                            textAlign: TextAlign.center,
-                            style: GoogleFonts.comfortaa(
-                              fontSize: 32,
-                              textStyle: const TextStyle(
-                                  letterSpacing: 0,
-                                  fontWeight: FontWeight.bold,
-                                  color: Colors.white),
-                            )),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.only(
-                            left: 15,
-                            right: 15,
-                            top: 10,
-                            bottom: 20), //apply padding to all four sides
-                        child: RichText(
-                          text: TextSpan(
-                            text: responseText1,
-                            style: GoogleFonts.comfortaa(
-                              fontSize: 18,
-                              textStyle: const TextStyle(
-                                  letterSpacing: 0,
-                                  fontWeight: FontWeight.bold,
-                                  color: Colors.white),
-                            ),
-                            children: <TextSpan>[
-                              TextSpan(
-                                text: responseText2,
-                                style: GoogleFonts.comfortaa(
-                                  fontSize: 18,
-                                  textStyle: TextStyle(
-                                      letterSpacing: 0,
-                                      fontWeight: FontWeight.bold,
-                                      color: responseColor == "Colors.green"
-                                          ? Colors.green
-                                          : Colors.red),
-                                ),
-                              ),
-                              TextSpan(
-                                text: responseText3,
-                                style: GoogleFonts.comfortaa(
-                                  fontSize: 18,
-                                  textStyle: const TextStyle(
-                                      letterSpacing: 0,
-                                      fontWeight: FontWeight.bold,
-                                      color: Colors.white),
-                                ),
-                              )
-                            ],
-                          ),
-                          textAlign: TextAlign.center,
-                        ),
-                      ),
-                      const Divider(
-                        color: Colors.white,
-                        thickness: 0.2,
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.only(
-                            left: 15,
-                            right: 15,
-                            top: 15,
-                            bottom: 15), //apply padding to all four sides
-                        child: SizedBox(
-                          width: 235, // <-- Your width
-                          height: 60, // <-- Your height
-                          child: ElevatedButton(
-                            style: ElevatedButton.styleFrom(
-                              primary: Colors.white,
-                              onPrimary: Colors.indigo,
-
-                              textStyle: TextStyle(fontSize: 20),
-                              shape: RoundedRectangleBorder(
-                                  borderRadius: BorderRadius.circular(32.0)),
-                              //minimumSize: const Size(40, 60),
-                            ),
-                            child: Row(
-                              crossAxisAlignment: CrossAxisAlignment.center,
-                              mainAxisAlignment: MainAxisAlignment.center,
-                              children: const [
-                                Icon(Icons.exit_to_app_outlined, size: 28),
-                                SizedBox(width: 10),
-                                Text("Powrót"),
-                              ],
-                            ),
-                            onPressed: () => Navigator.pop(
-                                context, _LoaderDialog.currentContext),
-                          ),
-                        ),
-                      )
-                    ],
-                  ),
-                ),
-              ),
-            ),
-          ),
-        );
-      },
-    );
-  }
-}
 
 class AddImage extends StatefulWidget {
   const AddImage({Key? key}) : super(key: key);
@@ -179,20 +44,15 @@ class _AddImageState extends State<AddImage> {
         final image = await ImagePicker()
             .pickImage(source: source, maxWidth: 400, maxHeight: 400);
         if (image == null) return;
-        bool isThePhotoFormatGood = false;
-        if (path.extension(path.basename(image.path)) == ".jpg" ||
-            path.extension(path.basename(image.path)) == ".jpeg" ||
-            path.extension(path.basename(image.path)) == ".png") {
-          isThePhotoFormatGood = true;
-        }
-        if (!validateFileExtension(image) || !isThePhotoFormatGood) {
-          //TODO Make a popcard communicating that GIFs are not allowed.
+
+        if (!validateFileExtension(image)) {
           responseTitle = "Wybrano niepoprawyny format";
           responseText1 = "Rozszerzenie twojego zdjęcia jest ";
           responseText2 = "niepoprawne";
           responseText3 = ". Akceptowane formaty : jpg, jpeg, png";
           responseColor = "Colors.red";
-          LoaderDialog.showLoadingDialog(context, _LoaderDialog);
+          LoaderDialog.showLoadingDialog(context, _LoaderDialog, responseTitle,
+              responseText1, responseText2, responseText3, responseColor);
           return;
         }
         final imageTemporary = await image.readAsBytes();
@@ -211,19 +71,15 @@ class _AddImageState extends State<AddImage> {
         final image = await ImagePicker()
             .pickImage(source: source, maxWidth: 400, maxHeight: 400);
         if (image == null) return;
-        bool isThePhotoFormatGood = false;
-        if (path.extension(path.basename(image.path)) == ".jpg" ||
-            path.extension(path.basename(image.path)) == ".jpeg" ||
-            path.extension(path.basename(image.path)) == ".png") {
-          isThePhotoFormatGood = true;
-        }
-        if (!validateFileExtension(image) || !isThePhotoFormatGood) {
+
+        if (!validateFileExtension(image)) {
           responseTitle = "Wybrano niepoprawyny format";
           responseText1 = "Rozszerzenie twojego zdjęcia jest ";
           responseText2 = "niepoprawne";
           responseText3 = ". Akceptowane formaty : jpg, jpeg, png";
           responseColor = "Colors.red";
-          LoaderDialog.showLoadingDialog(context, _LoaderDialog);
+          LoaderDialog.showLoadingDialog(context, _LoaderDialog, responseTitle,
+              responseText1, responseText2, responseText3, responseColor);
           return;
         }
         final imageTemporary = File(image.path);
@@ -328,7 +184,8 @@ class _AddImageState extends State<AddImage> {
         responseText3 = "odebrane, niewłaściwy adres serwera !";
         responseColor = "Colors.red";
       }
-      LoaderDialog.showLoadingDialog(context, _LoaderDialog);
+      LoaderDialog.showLoadingDialog(context, _LoaderDialog, responseTitle,
+          responseText1, responseText2, responseText3, responseColor);
     } on PlatformException catch (e) {
       if (kDebugMode) {
         print('Failed to send to server: $e');
@@ -373,7 +230,7 @@ class _AddImageState extends State<AddImage> {
                             fit: BoxFit.cover,
                           )
                         : Image.file(
-                  globals.mobileImageAdd!,
+                            globals.mobileImageAdd!,
                             width: 200,
                             height: 200,
                             fit: BoxFit.cover,
